refactor(qbittorrent-e2e): extract poller module
