+++ src/console/ci/qbittorrent/mod.rs
@@ -1,5 +1,6 @@
 pub mod bencode;
 pub mod client_role;
+pub mod poller;
 pub mod qbittorrent_client;
 pub mod runner;
 pub mod torrent_artifacts;
